==> .local/share/jupyter/nbgrader_cache/data301a/chewy06+project+2016-03-15 19:29:56 UTC/01-Introduction.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

import numpy as np
import pandas as pd
import os

# Hillary's emails

## Abstract:

In 2015 a controversy arose when it became publicily known that Hillary clinton used family's private emails server for official email conmmunications while she was the United States Secretary of State. Part of the controversy was the deletion of nearly 32,000 email that she deemed private. Nearly 2,100 emails have been marked classified by the state Departmet.Another 65 emiails were deemed secret and 22 deemed top secret. 
* OBJECTIVES
    - Dig through the 2,100 emails a create a statical model that differentias classified with normal emails. 
    - Determine the difference between classified email and deemed secret emails. 

## Dataset description

The original transcripts were transfered as pdf but kaggle broke it down nicely into manageble csv. The dataset is broken down into 5 tables each having pieces of information that can be found on the email data table. The data has been timestamped since the controversy process. The data is considered to be messy and imprecise. 
* [Link to Dataset](https://www.kaggle.com/kaggle/hillary-clinton-emails)

## Index

### [02-DataCleaning](02-DataCleaning.ipynb)

For my first notebook I focused on having separate tables that can be easily used to find data. The table names should tell a little of the type of content they carry. 

### [03-Exploration.ipynb](03-Exploration.ipynb)

On my exploraiton notebook I focused on finding out what content I was working with. I worked on merging and joining tables to accomplish certain tasks. Also method on how to access data much more efficiently. 

### [04-Modeling](04-Modeling.ipynb)

To find new ralationship between emails I decided to use Markov chain. The subject headings differ from Freedom of Information Act(FOIA) and original emails, therefore I decided to get a feel of the raw text based on Markov chain. 

###  [05-MachineLearning](05-MachineLearning.ipynb)

For machine learning I collected and parse the data based on the top people that contributed to the Hillary conversation. Could find a machine learning model to work with so outcome was not so great.
