==> .local/share/jupyter/nbgrader_cache/data301a/chewy06+project+2016-03-15 19:29:56 UTC/02-DataCleaning.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

import numpy as np
import pandas as pd
import os

# Data Cleaning

## Data

Read email data csv form data directory. Set the Id column to represent the table index.

emails=pd.read_csv('/data/chewy06/output/Emails.csv',index_col='Id')

Seperating data into two different tables FOIA and Extracted emails. One table will represent Email information the other email information from FOIA.

emailFOIA = emails[['DocNumber','MetadataSubject','MetadataTo','MetadataFrom','SenderPersonId','MetadataDateSent',
                    'MetadataDateReleased','MetadataPdfLink','MetadataCaseNumber','MetadataDocumentClass']]

emailExtracted = emails[['ExtractedSubject','ExtractedTo','ExtractedFrom','ExtractedCc','ExtractedDateSent',
                         'ExtractedCaseNumber', 'ExtractedDocNumber','ExtractedDateReleased','ExtractedReleaseInPartOrFull',
                         'ExtractedBodyText','RawText']]

emailFOIA.to_csv('/data/chewy06/output/EmailFOIA.csv',encoding='utf-8')

emailExtracted.to_csv('/data/chewy06/output/EmailExtracted.csv',encoding='utf-8')

FOIA=pd.read_csv('/data/chewy06/output/EmailFOIA.csv')

Extracted=pd.read_csv('/data/chewy06/output/EmailExtracted.csv')

assert list(FOIA.columns) == ['Id','DocNumber','MetadataSubject','MetadataTo','MetadataFrom','SenderPersonId',
                              'MetadataDateSent', 'MetadataDateReleased','MetadataPdfLink','MetadataCaseNumber',
                              'MetadataDocumentClass']

assert list(Extracted.columns) == ['Id','ExtractedSubject','ExtractedTo','ExtractedFrom','ExtractedCc','ExtractedDateSent',
                                        'ExtractedCaseNumber','ExtractedDocNumber','ExtractedDateReleased',
                                        'ExtractedReleaseInPartOrFull','ExtractedBodyText','RawText']

Read person.csv data from directory. Setting the Id to represent the table index. 

persons=pd.read_csv('/data/chewy06/output/Persons.csv', index_col='Id')

persons.head()

Read Aliases.csv data from directory. Setting the Id to represent the table index.

alias=pd.read_csv('/data/chewy06/output/Aliases.csv', index_col='Id')

alias.head()

Read Recipians.csv data from directory. Setting the Id to represent the table index.

recipians=pd.read_csv('/data/chewy06/output/EmailReceivers.csv', index_col='Id')

recipians.head()
