==> .local/share/jupyter/nbgrader_cache/data301a/chewy06+project+2016-03-15 19:29:56 UTC/04-Modeling.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

import numpy as np
import pandas as pd
import os

import types
from itertools import islice
import random

## Markov Chain on Hillary Clinton Emails

Markov Chain will used to build Classified email data based Confidential emails found. 

emails=pd.read_csv('/data/chewy06/output/Emails.csv', index_col='Id')
FOIA=pd.read_csv('/data/chewy06/output/EmailFOIA.csv', index_col='Id')
Extracted=pd.read_csv('/data/chewy06/output/EmailExtracted.csv', index_col='Id')
persons=pd.read_csv('/data/chewy06/output/Persons.csv', index_col='Id')
alias=pd.read_csv('/data/chewy06/output/Aliases.csv', index_col='Id')
recipians=pd.read_csv('/data/chewy06/output/EmailReceivers.csv', index_col='Id')

## Building the n-grams

Function that will return n-grams from input sequence. 

def build_ngrams(itr, n=2):
    for i in range(len(itr)-(n-1)):
        yield tuple(islice(itr, i, n + i))

## Building the markov chain

Building a function to return a dictionary based on a sequence of n-grams

def build_chain(ngrams, chain=None):
    if chain is None:
        chain = {}
    
    for ele in ngrams:
        n = len(ele)-1
        key = ele[:n:]
        value = ele[n::]
        val = min(value)
        if key in chain:
            chain[key].append(val)
        else:
            chain[key] = [val]
    return chain

Generating the Sequences for the Markov Chain

def generate_sequence(chain, m):
    random_key = random.choice(list(chain.keys()))
    tuple_size = len(random_key) 
    seq = list(random_key)
    
    for index in range(m - tuple_size):
        val = random.choice(chain[random_key])
        seq.append(val)
        random_key = tuple(seq[len(seq)-tuple_size::])
        
    return seq

import textwrap
import quicktoken

Extracting data from that contains confidential. Once the data has been collected n-grams will generated sequences based on the most common words used. By noticing the words being generated by the markov chain I cab have a sense of the what the subject or topic that was being addressed in the emails. 

confidential=emails[emails.MetadataSubject=='CONFIDENTIAL']

notConfidential=emails[emails.MetadataSubject=='RE:'] 

with open('confidential_emails.txt', 'w') as f:
    for text in confidential['RawText']:
        f.write(text.replace('\r\n', '\n'))
        f.write('\n')

with open('notConfidential_emails.txt', 'w') as f:
    for text in notConfidential['RawText']:
        f.write(text.replace('\r\n', '\n'))
        f.write('\n')

Function to generated probable email topics addressed. 

get_chain = quicktoken.files_to_lines(['confidential_emails.txt'])
seq = quicktoken.tokenize_lines(get_chain)
conf_chain = build_chain(build_ngrams(list(seq), n=3))

get_chain = quicktoken.files_to_lines(['notConfidential_emails.txt'])
seq = quicktoken.tokenize_lines(get_chain)
notConf_chain = build_chain(build_ngrams(list(seq), n=3))

def genEmailText(chain):
    s=''
    strList=generate_sequence(chain, 30)
    print(i+1,'.')
    for word in strList:
        s+=word+' '
    print(s)
    s=''
    print('\n')

Generating sequence for both confidential email as well. The diffrence in text will provide a sense of what the body text topic was. 

for i in range(3):
    genEmailText(conf_chain)

for i in range(3):
    genEmailText(notConf_chain)
