==> .local/share/jupyter/nbgrader_cache/data301a/chewy06+project+2016-03-15 19:29:56 UTC/05-MachineLearning.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

import numpy as np
import pandas as pd
import os

import types
from itertools import islice
import random

## Machine Learning

Hillary Clinton Dataset.

emails=pd.read_csv('/data/chewy06/output/Emails.csv', index_col='Id')
FOIA=pd.read_csv('/data/chewy06/output/EmailFOIA.csv', index_col='Id')
Extracted=pd.read_csv('/data/chewy06/output/EmailExtracted.csv', index_col='Id')
persons=pd.read_csv('/data/chewy06/output/Persons.csv', index_col='Id')
alias=pd.read_csv('/data/chewy06/output/Aliases.csv', index_col='Id')
recipians=pd.read_csv('/data/chewy06/output/EmailReceivers.csv', index_col='Id')

Loading top emial receivers. CSV created in previous notebooks. 

topEmailContributers=pd.read_csv('/data/chewy06/output/orderNames.csv', index_col='Id')

topEmailContributers[:10]

Created new dataset based on the name of person that sent the email. 

mergeWithPersonName=pd.merge(emails,persons,right_index=True,left_on=['SenderPersonId'])

Function to create files of emials by especific person. 

def createFile(fileName,data):
    with open(fileName, 'w') as f:
        for text in data['RawText']:
            f.write(text.replace('\r\n', '\n'))
            f.write('\n')
    

for name in topEmailContributers[:10]['Name']:
    data=mergeWithPersonName[mergeWithPersonName.Name==name]
    createFile(name+'.txt',data)

for name in topEmailContributers[:2]['Name']:
    with open(name+'.txt', 'r') as f:
        temp=f.read()
    temp=pd.Series(temp.lower().split(' '))
    words=pd.DataFrame({name:temp,
                        'id': pd.Series(range(temp.size))})
    words.index_col='id'
    print(name,': ', words[:10][name])
